refactor: convert some integration tests to jest unit tests
